pi4/requirements.txt
cassandra-driver
tqdm